Refact: refazendo algumas boas práticas

+++ src/controllers/Poligono.controller.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from "express";
 import { PoligonoService } from "../services/Poligono.service"
 import jwt from "jsonwebtoken"
+import { tokenNotFoundError } from "../errors/token-not-found-error";
+import { tokenInvalidError } from "../errors/token-invalid-error";
 
 export class PoligonoController {
 
@@ -8,21 +10,30 @@ export class PoligonoController {
         const token = req.headers.authorization
 
         if (!token) {
-            return res.status(401).json({ message: 'Token não encontrado' });
+            return res.status(401).json({ message: new tokenNotFoundError().message });
         }
 
-        jwt.verify(token, 'auth', (err: any, decoded: any) => {
-            if (err) {
-                return res.status(403).json({ message: 'Token inválido' });
-            }
+        const jwtError = await jwt.verify(token, 'auth', (err: any, decoded: any) => {
+            return err
         });
 
-        const service = new PoligonoService();
+        if (jwtError != null)
+            return res.status(403).json({ message: new tokenInvalidError().message });
+
+        const service = new PoligonoService()
 
         try {
-            return await service.create(req, res);
-        } catch (err) {
-            return res.status(400).json(err);
+            const result = await service.create(req, res);
+
+            return res.status(201).json({ message: 'Success', result: result })
+        } catch (err: any) {
+            let statusCode = 500;
+
+            if (err.statusCode) {
+                statusCode = err.statusCode;
+            }
+
+            return res.status(statusCode).json({ error: 'Get Poligon failed', message: err.message });
         }
     }
 
@@ -30,21 +41,30 @@ export class PoligonoController {
         const token = req.headers.authorization
 
         if (!token) {
-            return res.status(401).json({ message: 'Token não encontrado' });
+            return res.status(401).json({ message: new tokenNotFoundError().message });
         }
 
-        jwt.verify(token, 'auth', (err: any, decoded: any) => {
-            if (err) {
-                return res.status(403).json({ message: 'Token inválido' });
-            }
+        const jwtError = await jwt.verify(token, 'auth', (err: any, decoded: any) => {
+            return err
         });
 
-        const service = new PoligonoService();
+        if (jwtError != null)
+            return res.status(403).json({ message: new tokenInvalidError().message });
+
+        const service = new PoligonoService()
 
         try {
-            return await service.get(req, res);
-        } catch (err) {
-            return res.status(400).json(err)
+            const result = await service.get(req, res);
+
+            return res.status(200).json({ message: 'Success', result: result })
+        } catch (err: any) {
+            let statusCode = 500;
+
+            if (err.statusCode) {
+                statusCode = err.statusCode;
+            }
+
+            return res.status(statusCode).json({ error: 'Get Poligon failed', message: err.message });
         }
     }
 
@@ -52,21 +72,29 @@ export class PoligonoController {
         const token = req.headers.authorization
 
         if (!token) {
-            return res.status(401).json({ message: 'Token não encontrado' });
+            return res.status(401).json({ message: new tokenNotFoundError().message });
         }
 
-        jwt.verify(token, 'auth', (err: any, decoded: any) => {
-            if (err) {
-                return res.status(403).json({ message: 'Token inválido' });
-            }
+        const jwtError = await jwt.verify(token, 'auth', (err: any, decoded: any) => {
+            return err
         });
 
-        const service = new PoligonoService();
+        if (jwtError != null)
+            return res.status(403).json({ message: new tokenInvalidError().message });
 
+        const service = new PoligonoService()
         try {
-            return await service.getId(req, res);
-        } catch (err) {
-            return res.status(400).json(err);
+            const result = await service.getId(req, res);
+
+            return res.status(200).json({ message: 'Success', result: result })
+        } catch (err: any) {
+            let statusCode = 500;
+
+            if (err.statusCode) {
+                statusCode = err.statusCode;
+            }
+
+            return res.status(statusCode).json({ error: 'Get Poligon failed', message: err.message });
         }
     }
 
@@ -74,21 +102,29 @@ export class PoligonoController {
         const token = req.headers.authorization
 
         if (!token) {
-            return res.status(401).json({ message: 'Token não encontrado' });
+            return res.status(401).json({ message: new tokenNotFoundError().message });
         }
 
-        jwt.verify(token, 'auth', (err: any, decoded: any) => {
-            if (err) {
-                return res.status(403).json({ message: 'Token inválido' });
-            }
+        const jwtError = await jwt.verify(token, 'auth', (err: any, decoded: any) => {
+            return err
         });
 
-        const service = new PoligonoService();
+        if (jwtError != null)
+            return res.status(403).json({ message: new tokenInvalidError().message });
 
+        const service = new PoligonoService()
         try {
-            return await service.delete(req, res)
-        } catch (err) {
-            return res.status(400).json(err);
+            const result = await service.delete(req, res)
+
+            return res.status(204).json({ message: 'Success'})
+        } catch (err: any) {
+            let statusCode = 500;
+
+            if (err.statusCode) {
+                statusCode = err.statusCode;
+            }
+
+            return res.status(statusCode).json({ error: 'Get Poligon failed', message: err.message });
         }
     }
 
@@ -96,21 +132,29 @@ export class PoligonoController {
         const token = req.headers.authorization
 
         if (!token) {
-            return res.status(401).json({ message: 'Token não encontrado' });
+            return res.status(401).json({ message: new tokenNotFoundError().message });
         }
 
-        jwt.verify(token, 'auth', (err: any, decoded: any) => {
-            if (err) {
-                return res.status(403).json({ message: 'Token inválido' });
-            }
+        const jwtError = await jwt.verify(token, 'auth', (err: any, decoded: any) => {
+            return err
         });
 
-        const service = new PoligonoService();
+        if (jwtError != null)
+            return res.status(403).json({ message: new tokenInvalidError().message });
 
+        const service = new PoligonoService()
         try {
-            return await service.put(req, res)
-        } catch (err) {
-            return res.status(400).json(err);
+            const result = await service.put(req, res)
+
+            return res.status(200).json({ message: 'Success', result: result })
+        } catch (err: any) {
+            let statusCode = 500;
+
+            if (err.statusCode) {
+                statusCode = err.statusCode;
+            }
+
+            return res.status(statusCode).json({ error: 'Get Poligon failed', message: err.message });
         }
     }
 }
